--- a/helpers.go
+++ b/helpers.go
@@ -177,15 +177,5 @@ func toBytes32(x []byte) [32]byte {
 
 // hash defines a function that returns the sha256 hash of the data passed in.
 func hash(data []byte) [32]byte {
-	var hash [32]byte
-
-	h := sha256.New()
-	// The hash interface never returns an error, for that reason
-	// we are not handling the error below. For reference, it is
-	// stated here https://golang.org/pkg/hash/#Hash
-	// #nosec G104
-	h.Write(data)
-	h.Sum(hash[:0])
-
-	return hash
+	return sha256.Sum256(data)
 }