Before:
```
2019/08/02 12:27:01 324.9995ms to unmarshal
2019/08/02 12:27:01 85.0004ms to marshal
2019/08/02 12:27:01 profile: cpu profiling enabled, Z:\Temp\profile279911603\cpu.pprof
2019/08/02 12:27:02 466.1326ms hash tree root
2019/08/02 12:27:02 90.9946ms subsequent hash tree root
2019/08/02 12:27:02 profile: cpu profiling disabled, Z:\Temp\profile279911603\cpu.pprof
```

After:
```
2019/08/02 12:27:11 319.9992ms to unmarshal
2019/08/02 12:27:11 38.0001ms to marshal
2019/08/02 12:27:11 profile: cpu profiling enabled, Z:\Temp\profile562274255\cpu.pprof
2019/08/02 12:27:11 252.9953ms hash tree root
2019/08/02 12:27:11 44.0049ms subsequent hash tree root
2019/08/02 12:27:12 profile: cpu profiling disabled, Z:\Temp\profile562274255\cpu.pprof
```