@rauljordan 2019-08-02
Hi @egonelbre could you add this to the end of `deps.bzl`

```
_maybe(
        go_repository,
        name = "com_github_minio_sha256_simd",
        commit = "05b4dd3047e5d6e86cb4e0477164b850cd896261",
        importpath = "github.com/minio/sha256-simd",
    )
```

and then run `bazel run //:gazelle` ? 

Thanks for the contributions - the changes are awesome!